15.12.22

+++ admin/cdethquest/op/equipment.php
@@ -11,8 +11,8 @@ if (isset($_POST['filter'])){
 }
 $regspag = 20;
 ?>
-  <p><?php echo HelpEquipment; ?></p>
-  <p><?php echo Active_language; ?> <b><?php echo $site->activeLanguage(); ?></b></p>
+<p><?php echo HelpEquipment; ?></p>
+<p><?php echo Active_language; ?> <b><?php echo $site->activeLanguage(); ?></b></p>
 <?php
 if (!isset($_POST['item'])){
 ?>
@@ -83,58 +83,71 @@ if (!isset($_POST['item'])){
   }
 }else{
   $item = new Equipment($site->getDatalink(), $_POST['item']);
-  if (isset($_POST['name'])
-  and isset($_POST['price'])
-  and isset($_POST['premium'])
-  and isset($_POST['description'])){
-    $item->setName($_POST['name']);
-    $item->setPrice($_POST['price']);
-    $item->setPremium($_POST['premium']);
-    $item->setDescription($_POST['description']);
-    if (isset($_POST['permanent'])){
-      $item->setPermanent(true);
+  if( isset( $_POST['name'] )
+  and isset( $_POST['price'] )
+  and isset( $_POST['premium'] )
+  and isset( $_POST['description'] ) 
+  and isset( $_POST['health'] ) 
+  and isset( $_POST['speed'] ) 
+  and isset( $_POST['strength'] ) 
+  and isset( $_POST['dexterity'] ) 
+  and isset( $_POST['constitution'] ) 
+  and isset( $_POST['intelligence'] ) ){
+    $item->setName( $_POST['name'] );
+    $item->setPrice( $_POST['price'] );
+    $item->setPremium( $_POST['premium'] );
+    $item->setDescription( $_POST['description'] );
+    $item->setMaxhealth( $_POST['health'] );
+    $item->setSpeed( $_POST['speed'] );
+    $item->setStrength( $_POST['strength'] );
+    $item->setDexterity( $_POST['dexterity'] );
+    $item->setConstitution( $_POST['constitution'] );
+    $item->setIntelligence( $_POST['intelligence'] );
+    if( isset( $_POST['permanent'] ) ){
+      $item->setPermanent( true );
     }else{
-      $item->setPermanent(false);
+      $item->setPermanent( false );
     }
-    if (isset($_FILES['icon'])){
-      $item->setIcon($site->imageUpload($_FILES['icon'], $module->getFolder()));
+    if( isset( $_FILES['icon'] ) ){
+      $item->setIcon( $site->imageUpload( $_FILES['icon'], $module->getFolder() ) );
     }
-    if (isset($_FILES['maleimage'])){
-      $item->setMaleimage($site->imageUpload($_FILES['maleimage'], $module->getFolder()));
+    if( isset( $_FILES['maleimage'] ) ){
+      $item->setMaleimage( $site->imageUpload( $_FILES['maleimage'], $module->getFolder() ) );
     }
-    if (isset($_FILES['femaleimage'])){
-      $item->setFemaleimage($site->imageUpload($_FILES['femaleimage'], $module->getFolder()));
+    if( isset( $_FILES['femaleimage'] ) ){
+      $item->setFemaleimage( $site->imageUpload( $_FILES['femaleimage'], $module->getFolder() ) );
     }
     if( isset( $_POST['forsale'] ) ){
       $item->setForsale( true );
     }else{
       $item->setForsale( false );
     }
-    if (isset($_POST['delete'])){
+    if( isset( $_POST['deleteicon'] ) 
+    or isset( $_POST['delete'] ) ){
+      if( file_exists( $site->getRoot().'uploads/'.$module->getFolder().'/'.$item->getIcon() ) ){
+        unlink( $site->getRoot().'uploads/'.$module->getFolder().'/'.$item->getIcon() );
+      }
+      $item->setIcon( '' );
+      $item->save();
+    }
+    if( isset( $_POST['delete'] ) ){
       echo $item->delete();
     }else{
       echo $item->save();
     }
   }
-  if (isset($_POST['deleteicon'])){
-    if(file_exists($site->getRoot().'uploads/'.$module->getFolder().'/'.$item->getIcon())){
-      unlink($site->getRoot().'uploads/'.$module->getFolder().'/'.$item->getIcon());
+  if( isset( $_POST['deletemaleimage'] ) ){
+    if( file_exists( $site->getRoot().'uploads/'.$module->getFolder().'/'.$item->getMaleimage() ) ){
+      unlink( $site->getRoot().'uploads/'.$module->getFolder().'/'.$item->getMaleimage() );
     }
-    $item->setIcon('');
+    $item->setMaleimage( '' );
     $item->save();
   }
-  if (isset($_POST['deletemaleimage'])){
-    if(file_exists($site->getRoot().'uploads/'.$module->getFolder().'/'.$item->getMaleimage())){
-      unlink($site->getRoot().'uploads/'.$module->getFolder().'/'.$item->getMaleimage());
+  if( isset( $_POST['deletefemaleimage'] ) ){
+    if( file_exists( $site->getRoot().'uploads/'.$module->getFolder().'/'.$item->getFemaleimage() ) ){
+      unlink( $site->getRoot().'uploads/'.$module->getFolder().'/'.$item->getFemaleimage() );
     }
-    $item->setMaleimage('');
-    $item->save();
-  }
-  if (isset($_POST['deletefemaleimage'])){
-    if(file_exists($site->getRoot().'uploads/'.$module->getFolder().'/'.$item->getFemaleimage())){
-      unlink($site->getRoot().'uploads/'.$module->getFolder().'/'.$item->getFemaleimage());
-    }
-    $item->setFemaleimage('');
+    $item->setFemaleimage( '' );
     $item->save();
   }
   ?>
@@ -149,30 +162,50 @@ if (!isset($_POST['item'])){
     <input type="hidden" name="pag" value="<?php echo $pag; ?>">
     <input type="hidden" name="item" value="<?php echo $item->getId(); ?>">
     <div class="edblock">
+      <p class="formimg">
+        <?php
+        if( $item->getIcon() == '' ){
+          echo Item_image.'<br>
+          <input type="file" name="icon">';
+        }else{
+          echo Item_image.'<br>
+          <img src="'.$_GET['root'].'uploads/'.$module->getFolder().'/'.$item->getIcon().'"><br>
+          <input type="checkbox" name="deleteicon" value="1"> '.Delete_picture;
+        }
+        ?>
+      </p>
       <p class="pinput"><?php echo Item_name; ?><br>
       <input type="text" name="name" value="<?php echo $item->getName(); ?>"></p>
       <p class="pinput"><?php echo Item_price; ?><br>
       <input type="number" name="price" value="<?php echo $item->getPrice(); ?>" style="width:6em"></p>
       <p class="pinput"><?php echo Premium_price; ?><br>
       <input type="number" name="premium" value="<?php echo $item->getPremium(); ?>" style="width:6em"></p>
+      <p><?php echo Item_description; ?><br>
+      <textarea name="description"><?php echo $item->getDescription(); ?></textarea></p>
       <p class="pinput"><input type="checkbox" name="permanent" <?php
-      if ($item->getPermanent()){
+      if( $item->getPermanent() ){
         echo ' checked';
       }
       ?>> <?php echo Permanent; ?></p>
       <p class="pinput"><input type="checkbox" name="forsale" value="1"<?php if( $item->getForsale() ){ echo ' checked'; } ?>> <?php echo For_sale; ?></p>
     </div>
     <div class="edblock">
-      <p><?php echo Item_description; ?><br>
-      <textarea name="description"><?php echo $item->getDescription(); ?></textarea></p>
+      <p class="pinput"><?php echo Health_boost ?>:<br>
+      <input type="number" name="health" value="<?php echo $item->getMaxhealth(); ?>" style="width:6em"></p>
+      <p class="pinput"><?php echo Speed_boost ?>:<br>
+      <input type="number" name="speed" value="<?php echo $item->getSpeed(); ?>" style="width:6em"></p>
+      <p class="pinput"><?php echo Strength_boost ?>:<br>
+      <input type="number" name="strength" value="<?php echo $item->getStrength(); ?>" style="width:6em"></p>
+      <p class="pinput"><?php echo Dexterity_boost ?>:<br>
+      <input type="number" name="dexterity" value="<?php echo $item->getDexterity(); ?>" style="width:6em"></p>
+      <p class="pinput"><?php echo Constitution_boost ?>:<br>
+      <input type="number" name="constitution" value="<?php echo $item->getConstitution(); ?>" style="width:6em"></p>
+      <p class="pinput"><?php echo Intelligence_boost ?>:<br>
+      <input type="number" name="intelligence" value="<?php echo $item->getIntelligence(); ?>" style="width:6em"></p>
     </div>
     <p><input type="submit" value="<?php echo Save_item; ?>">
     <input type="checkbox" name="delete" value="1"><?php echo Delete_item; ?></p>
     <?php 
-    if ($item->getIcon() == ''){
-      echo '<div class="formimg">'.Item_image.'<br>
-      <input type="file" name="icon"></div>';
-    }
     if ($item->getMaleimage() == ''){
       echo '<div class="formimg pinput">'.Item_male_image.':<br>
       <div class="equipment-preview" style="background-image:url('.$_GET['root'].'admin/'.$module->getFolder().'/css/images/generic-male.png)">
@@ -190,18 +223,6 @@ if (!isset($_POST['item'])){
     ?>
   </form>
   <?php
-  if ($item->getIcon() != ''){
-    echo '<form name="icon" onsubmit="event.preventDefault(); backend.post(this);" method="post" action="">
-    <input type="hidden" name="filter" value="'.$filter.'">
-    <input type="hidden" name="pag" value="'.$pag.'">
-    <input type="hidden" name="item" value="'.$item->getId().'">
-    <input type="hidden" name="deleteicon" value="1">
-    <div class="formimg">
-    <img src="'.$_GET['root'].'uploads/'.$module->getFolder().'/'.$item->getIcon().'"><br>
-    <input type="submit" value="'.Delete_picture.'">
-    </div>
-    </form>';
-  }
   if ($item->getMaleimage() != ''){
     echo '<form name="maleimage" onsubmit="event.preventDefault(); backend.post(this);" method="post" action="">
     <input type="hidden" name="filter" value="'.$filter.'">
